@kdjdbbfk 2024-03-28
arm-old
goos: linux
goarch: arm64
pkg: crypto/md5
BenchmarkHash8Bytes-4            	 6902066	       174 ns/op	  46.03 MB/s
BenchmarkHash1K-4                	  582655	      2059 ns/op	 497.33 MB/s
BenchmarkHash8K-4                	   78613	     15262 ns/op	 536.76 MB/s
BenchmarkHash8BytesUnaligned-4   	 6906164	       174 ns/op	  46.01 MB/s
BenchmarkHash1KUnaligned-4       	  582601	      2058 ns/op	 497.49 MB/s
BenchmarkHash8KUnaligned-4       	   78133	     15357 ns/op	 533.45 MB/s
PASS
ok  	crypto/md5	7.916s

arm-new
goos: linux
goarch: arm64
pkg: crypto/md5
BenchmarkHash8Bytes-4            	 6905114	       174 ns/op	  46.07 MB/s
BenchmarkHash1K-4                	  582960	      2054 ns/op	 498.63 MB/s
BenchmarkHash8K-4                	   78751	     15251 ns/op	 537.15 MB/s
BenchmarkHash8BytesUnaligned-4   	 6911649	       174 ns/op	  46.06 MB/s
BenchmarkHash1KUnaligned-4       	  581740	      2053 ns/op	 498.72 MB/s
BenchmarkHash8KUnaligned-4       	   78244	     15332 ns/op	 534.32 MB/s
PASS
ok  	crypto/md5	7.908s