@kdjdbbfk 2024-03-28
goos: linux
goarch: arm64
pkg: crypto/md5
                      │  .\old.txt  │             .\new.txt              │
                      │   sec/op    │   sec/op     vs base               │
Hash8Bytes-8            175.0n ± 2%   175.0n ± 1%       ~ (p=0.070 n=10)
Hash1K-8                2.065µ ± 0%   2.060µ ± 0%  -0.22% (p=0.041 n=10)
Hash8K-8                15.31µ ± 0%   15.29µ ± 0%  -0.11% (p=0.050 n=10)
Hash8BytesUnaligned-8   174.0n ± 1%   174.0n ± 1%       ~ (p=0.721 n=10)
Hash1KUnaligned-8       2.067µ ± 0%   2.059µ ± 0%  -0.41% (p=0.001 n=10)
Hash8KUnaligned-8       15.44µ ± 0%   15.45µ ± 0%       ~ (p=0.912 n=10)
geomean                 1.770µ        1.768µ       -0.11%

                      │  .\old.txt   │              .\new.txt              │
                      │     B/s      │     B/s       vs base               │
Hash8Bytes-8            43.58Mi ± 2%   43.69Mi ± 0%  +0.24% (p=0.022 n=10)
Hash1K-8                473.1Mi ± 0%   474.0Mi ± 0%  +0.20% (p=0.037 n=10)
Hash8K-8                510.4Mi ± 0%   511.0Mi ± 0%  +0.11% (p=0.045 n=10)
Hash8BytesUnaligned-8   43.80Mi ± 0%   43.82Mi ± 0%       ~ (p=0.170 n=10)
Hash1KUnaligned-8       472.5Mi ± 0%   474.3Mi ± 0%  +0.38% (p=0.001 n=10)
Hash8KUnaligned-8       506.1Mi ± 0%   505.8Mi ± 0%       ~ (p=0.912 n=10)
geomean                 219.0Mi        219.3Mi       +0.15%