@kdjdbbfk 2024-03-28
This optimization is more pronounced when parsing large files, for example, when parsing 3G iso, the running time is reduced from 8.65s to 7.39s, which is about 9% less time