@kdjdbbfk 2024-09-06
Resubmitted the PR at: https://github.com/golang/go/pull/69302